requirements.txt
torch
torchvision
transformers
datasets
numpy
pillow
flask
requests
